GraphNeuralNetworks

Documentation for GraphNeuralNetworks.

GraphNeuralNetworks.ChebConv

GraphNeuralNetworks.EdgeConv

GraphNeuralNetworks.FeaturedGraph

GraphNeuralNetworks.GATConv

GraphNeuralNetworks.GCNConv

GraphNeuralNetworks.GINConv

GraphNeuralNetworks.GatedGraphConv

GraphNeuralNetworks.GlobalPool

GraphNeuralNetworks.GraphConv

GraphNeuralNetworks.LocalPool

GraphNeuralNetworks.MessagePassing

GraphNeuralNetworks.TopKPool

GraphNeuralNetworks.add_self_loops

GraphNeuralNetworks.bypass_graph

GraphNeuralNetworks.edge_feature

GraphNeuralNetworks.edge_index

GraphNeuralNetworks.global_feature

GraphNeuralNetworks.graph

GraphNeuralNetworks.message

GraphNeuralNetworks.node_feature

GraphNeuralNetworks.normalized_laplacian

GraphNeuralNetworks.propagate

GraphNeuralNetworks.scaled_laplacian

GraphNeuralNetworks.update

GraphNeuralNetworks.ChebConv — Type

ChebConv(in=>out, k; bias=true, init=glorot_uniform)

Chebyshev spectral graph convolutional layer.

Arguments

in: The dimension of input features.

out: The dimension of output features.

k: The order of Chebyshev polynomial.

bias: Add learnable bias.

init: Weights' initializer.

GraphNeuralNetworks.EdgeConv — Type

EdgeConv(nn; aggr=max)

Edge convolutional layer.

Arguments

nn: A neural network (e.g. a Dense layer or a MLP).

aggr: An aggregate function applied to the result of message function. +, max and mean are available.

GraphNeuralNetworks.FeaturedGraph — Type

FeaturedGraph(g; [graph_type, dir, num_nodes, nf, ef, gf])
FeaturedGraph(fg::FeaturedGraph; [nf, ef, gf])

A type representing a graph structure and storing also arrays that contain features associated to nodes, edges, and the whole graph.

A FeaturedGraph can be constructed out of different objects g representing the connections inside the graph, while the internal representation type is governed by graph_type. When constructed from another featured graph fg, the internal graph representation is preserved and shared.

A FeaturedGraph is a LightGraphs' AbstractGraph, therefore any functionality from the LightGraphs' graph library can be used on it.

Arguments

g: Some data representing the graph topology. Possible type are

An adjacency matrix

An adjacency list.

A tuple containing the source and target vectors (COO representation)

A LightGraphs' graph.

graph_type: A keyword argument that specifies the underlying representation used by the FeaturedGraph. Currently supported values are

:coo. Graph represented as a tuple (source, target), such that the k-th edge connects the node source[k] to node target[k]. Optionally, also edge weights can be given: (source, target, weights).

:sparse. A sparse adjacency matrix representation.

:dense. A dense adjacency matrix representation.

Default :coo.

dir. The assumed edge direction when given adjacency matrix or adjacency list input data g. Possible values are :out and :in. Defaul :out.

num_nodes. The number of nodes. If not specified, inferred from g. Default nothing.

nf: Node features. Either nothing, or an array whose last dimension has size num_nodes. Default nothing.

ef: Edge features. Either nothing, or an array whose last dimension has size num_edges. Default nothing.

gf: Global features. Default nothing.

Usage.

using Flux, GraphNeuralNetworks

# Construct from adjacency list representation
g = [[2,3], [1,4,5], [1], [2,5], [2,4]]
fg = FeaturedGraph(g)

# Number of nodes and edges
fg.num_nodes  # 5
fg.num_edges  # 10 

# Same graph in COO representation
s = [1,1,2,2,2,3,4,4,5,5]
t = [2,3,1,4,5,3,2,5,2,4]
fg = FeaturedGraph((s, t))
fg = FeaturedGraph(s, t) # other convenience constructor

# From a LightGraphs' graph
fg = FeaturedGraph(erdos_renyi(100, 20))

# Copy featured graph while also adding node features
fg = FeaturedGraph(fg, nf=rand(100, 5))

# Send to gpu
fg = fg |> gpu

# Collect edges' source and target nodes.
# Both source and target are vectors of length num_edges
source, target = edge_index(fg)

See also graph, edge_index, node_feature, edge_feature, and global_feature

GraphNeuralNetworks.GATConv — Type

GATConv(in => out;
        heads=1,
        concat=true,
        init=glorot_uniform    
        bias=true, 
        negative_slope=0.2)

Graph attentional layer.

Arguments

in: The dimension of input features.

out: The dimension of output features.

bias::Bool: Keyword argument, whether to learn the additive bias.

heads: Number attention heads

concat: Concatenate layer output or not. If not, layer output is averaged over the heads.

negative_slope::Real: Keyword argument, the parameter of LeakyReLU.

GraphNeuralNetworks.GCNConv — Type

GCNConv(in => out, σ=identity; bias=true, init=glorot_uniform)

Graph convolutional layer.

Arguments

in: The dimension of input features.

out: The dimension of output features.

σ: Activation function.

bias: Add learnable bias.

init: Weights' initializer.

The input to the layer is a node feature array X of size (num_features, num_nodes).

GraphNeuralNetworks.GINConv — Type

GINConv(nn; eps = 0f0)

Graph Isomorphism Network.

Arguments

nn: A neural network/layer.

eps: Weighting factor.

The definition of this is as defined in the original paper, Xu et. al. (2018) https://arxiv.org/abs/1810.00826.

GraphNeuralNetworks.GatedGraphConv — Type

GatedGraphConv(out, num_layers; aggr=+, init=glorot_uniform)

Gated graph convolution layer.

Arguments

out: The dimension of output features.

num_layers: The number of gated recurrent unit.

aggr: An aggregate function applied to the result of message function. +, -,

*, /, max, min and mean are available.

GraphNeuralNetworks.GlobalPool — Type

GlobalPool(aggr, dim...)

Global pooling layer.

It pools all features with aggr operation.

Arguments

aggr: An aggregate function applied to pool all features.

GraphNeuralNetworks.GraphConv — Type

GraphConv(in => out, σ=identity, aggr=+; bias=true, init=glorot_uniform)

Graph neural network layer.

Arguments

in: The dimension of input features.

out: The dimension of output features.

σ: Activation function.

aggr: An aggregate function applied to the result of message function. +, -,

*, /, max, min and mean are available.

bias: Add learnable bias.

init: Weights' initializer.

GraphNeuralNetworks.LocalPool — Type

LocalPool(aggr, cluster)

Local pooling layer.

It pools features with aggr operation accroding to cluster. It is implemented with scatter operation.

Arguments

aggr: An aggregate function applied to pool all features.

cluster: An index structure which indicates what features to aggregate with.

GraphNeuralNetworks.MessagePassing — Type

MessagePassing

The abstract type from which all message passing layers are derived.

Related methods are propagate, message, update, update_edge, and update_global.

GraphNeuralNetworks.TopKPool — Type

TopKPool(adj, k, in_channel)

Top-k pooling layer.

Arguments

adj: Adjacency matrix of a graph.

k: Top-k nodes are selected to pool together.

in_channel: The dimension of input channel.

GraphNeuralNetworks.add_self_loops — Method

add_self_loops(fg::FeaturedGraph)

Return a featured graph with the same features as fg but also adding edges connecting the nodes to themselves.

Nodes with already existing self-loops will obtain a second self-loop.

GraphNeuralNetworks.bypass_graph — Function

bypass_graph(nf_func, ef_func, gf_func)

Bypassing graph in FeaturedGraph and let other layer process (node, edge and global)features only.

GraphNeuralNetworks.edge_feature — Method

edge_feature(fg::FeaturedGraph)

Return the edge features of fg.

GraphNeuralNetworks.edge_index — Method

edge_index(fg::FeaturedGraph)

Return a tuple containing two vectors, respectively storing the source and target nodes for each edges in fg.

s, t = edge_index(fg)

GraphNeuralNetworks.global_feature — Method

global_feature(fg::FeaturedGraph)

Return the global features of fg.

GraphNeuralNetworks.graph — Method

graph(fg::FeaturedGraph)

Return the underlying implementation of the graph structure of fg, either an adjacency matrix or an edge list in the COO format.

GraphNeuralNetworks.message — Function

message(mp::MessagePassing, x_i, x_j, [e_ij, u])

Message function for the message-passing scheme, returning the message from node j to node i . In the message-passing scheme. the incoming messages from the neighborhood of i will later be aggregated in order to update the features of node i.

By default, the function returns x_j. Layers subtyping MessagePassing should specialize this method with custom behavior.

Arguments

mp: A MessagePassing layer.

x_i: Features of the central node i.

x_j: Features of the neighbor j of node i.

e_ij: Features of edge (i, j).

u: Global features.

See also update and propagate.

GraphNeuralNetworks.node_feature — Method

node_feature(fg::FeaturedGraph)

Return the node features of fg.

GraphNeuralNetworks.normalized_laplacian — Function

normalized_laplacian(fg, T=Float32; add_self_loops=false, dir=:out)

Normalized Laplacian matrix of graph g.

Arguments

fg: A FeaturedGraph.

T: result element type.

add_self_loops: add self-loops while calculating the matrix.

dir: the edge directionality considered (:out, :in, :both).

GraphNeuralNetworks.propagate — Function

propagate(mp::MessagePassing, fg::FeaturedGraph, aggr)
propagate(mp::MessagePassing, fg::FeaturedGraph, E, X, u, aggr)

Perform the sequence of operation implementing the message-passing scheme and updating node, edge, and global features X, E, and u respectively.

The computation involved is the following:

M = compute_batch_message(mp, fg, E, X, u) 
E = update_edge(mp, M, E, u)
M̄ = aggregate_neighbors(mp, aggr, fg, M)
X = update(mp, M̄, X, u)
u = update_global(mp, E, X, u)

Custom layers sub-typing MessagePassing typically call define their own update and message function, than call this method in the forward pass:

function (l::GNNLayer)(fg, X)
    ... some prepocessing if needed ...
    E = nothing
    u = nothing
    propagate(l, fg, E, X, u, +)
end

See also message and update.

GraphNeuralNetworks.scaled_laplacian — Function

scaled_laplacian(fg, T=Float32; dir=:out)

Scaled Laplacian matrix of graph g, defined as $\hat{L} = \frac{2}{\lambda_{max}} L - I$ where $L$ is the normalized Laplacian matrix.

Arguments

fg: A FeaturedGraph.

T: result element type.

dir: the edge directionality considered (:out, :in, :both).

GraphNeuralNetworks.update — Function

update(mp::MessagePassing, m̄, x, [u])

Update function for the message-passing scheme, returning a new set of node features x′ based on old features x and the incoming message from the neighborhood aggregation m̄.

By default, the function returns m̄. Layers subtyping MessagePassing should specialize this method with custom behavior.

Arguments

mp: A MessagePassing layer.

m̄: Aggregated edge messages from the message function.

x: Node features to be updated.

u: Global features.

See also message and propagate.