Graphs

GraphNeuralNetworks.FeaturedGraph — Type

FeaturedGraph(g; [graph_type, dir, num_nodes, nf, ef, gf])
FeaturedGraph(fg::FeaturedGraph; [nf, ef, gf])

A type representing a graph structure and storing also arrays that contain features associated to nodes, edges, and the whole graph.

A FeaturedGraph can be constructed out of different objects g representing the connections inside the graph, while the internal representation type is governed by graph_type. When constructed from another featured graph fg, the internal graph representation is preserved and shared.

A FeaturedGraph is a LightGraphs' AbstractGraph, therefore any functionality from the LightGraphs' graph library can be used on it.

Arguments

g: Some data representing the graph topology. Possible type are

An adjacency matrix

An adjacency list.

A tuple containing the source and target vectors (COO representation)

A LightGraphs' graph.

graph_type: A keyword argument that specifies the underlying representation used by the FeaturedGraph. Currently supported values are

:coo. Graph represented as a tuple (source, target), such that the k-th edge connects the node source[k] to node target[k]. Optionally, also edge weights can be given: (source, target, weights).

:sparse. A sparse adjacency matrix representation.

:dense. A dense adjacency matrix representation.

Default :coo.

dir. The assumed edge direction when given adjacency matrix or adjacency list input data g. Possible values are :out and :in. Defaul :out.

num_nodes. The number of nodes. If not specified, inferred from g. Default nothing.

nf: Node features. Either nothing, or an array whose last dimension has size num_nodes. Default nothing.

ef: Edge features. Either nothing, or an array whose last dimension has size num_edges. Default nothing.

gf: Global features. Default nothing.

Usage.

using Flux, GraphNeuralNetworks

# Construct from adjacency list representation
g = [[2,3], [1,4,5], [1], [2,5], [2,4]]
fg = FeaturedGraph(g)

# Number of nodes and edges
fg.num_nodes  # 5
fg.num_edges  # 10 

# Same graph in COO representation
s = [1,1,2,2,2,3,4,4,5,5]
t = [2,3,1,4,5,3,2,5,2,4]
fg = FeaturedGraph((s, t))
fg = FeaturedGraph(s, t) # other convenience constructor

# From a LightGraphs' graph
fg = FeaturedGraph(erdos_renyi(100, 20))

# Copy featured graph while also adding node features
fg = FeaturedGraph(fg, nf=rand(100, 5))

# Send to gpu
fg = fg |> gpu

# Collect edges' source and target nodes.
# Both source and target are vectors of length num_edges
source, target = edge_index(fg)

See also graph, edge_index, node_feature, edge_feature, and global_feature

GraphNeuralNetworks.add_self_loops — Method

add_self_loops(fg::FeaturedGraph)

Return a featured graph with the same features as fg but also adding edges connecting the nodes to themselves.

Nodes with already existing self-loops will obtain a second self-loop.

GraphNeuralNetworks.edge_feature — Method

edge_feature(fg::FeaturedGraph)

Return the edge features of fg.

GraphNeuralNetworks.edge_index — Method

edge_index(fg::FeaturedGraph)

Return a tuple containing two vectors, respectively storing the source and target nodes for each edges in fg.

s, t = edge_index(fg)

GraphNeuralNetworks.global_feature — Method

global_feature(fg::FeaturedGraph)

Return the global features of fg.

GraphNeuralNetworks.node_feature — Method

node_feature(fg::FeaturedGraph)

Return the node features of fg.

GraphNeuralNetworks.normalized_laplacian — Function

normalized_laplacian(fg, T=Float32; add_self_loops=false, dir=:out)

Normalized Laplacian matrix of graph g.

Arguments

fg: A FeaturedGraph.

T: result element type.

add_self_loops: add self-loops while calculating the matrix.

dir: the edge directionality considered (:out, :in, :both).

GraphNeuralNetworks.scaled_laplacian — Function

scaled_laplacian(fg, T=Float32; dir=:out)

Scaled Laplacian matrix of graph g, defined as $\hat{L} = \frac{2}{\lambda_{max}} L - I$ where $L$ is the normalized Laplacian matrix.

Arguments

fg: A FeaturedGraph.

T: result element type.

dir: the edge directionality considered (:out, :in, :both).