Convolutional Layers

Index

GraphNeuralNetworks.CGConv

GraphNeuralNetworks.ChebConv

GraphNeuralNetworks.EdgeConv

GraphNeuralNetworks.GATConv

GraphNeuralNetworks.GCNConv

GraphNeuralNetworks.GINConv

GraphNeuralNetworks.GatedGraphConv

GraphNeuralNetworks.GraphConv

GraphNeuralNetworks.NNConv

GraphNeuralNetworks.ResGatedGraphConv

GraphNeuralNetworks.SAGEConv

Docs

GraphNeuralNetworks.CGConv — Type

CGConv((nin, ein) => nout, f, act=identity; bias=true, init=glorot_uniform, residual=false)

The crystal graph convolutional layer from the paper Crystal Graph Convolutional Neural Networks for an Accurate and Interpretable Prediction of Material Properties. Performs the operation

\[\mathbf{x}_i' = \mathbf{x}_i + \sum_{j\in N(i)}\sigma(W_f \mathbf{z}_{ij} + \mathbf{b}_f)\, act(W_s \mathbf{z}_{ij} + \mathbf{b}_s)\]

where $\mathbf{z}_{ij}$ is the node and edge features concatenation $[\mathbf{x}_i \| \mathbf{x}_j \| \mathbf{e}_{j\to i}]$ and $\sigma$ is the sigmoid function. The residual $\mathbf{x}_i$ is added only if residual=true and the output size is the same as the input size.

Arguments

nin: The dimension of input node features.

nout: The dimension of input edge features.

out: The dimension of output node features.

act: Activation function.

bias: Add learnable bias.

init: Weights' initializer.

residual: Add a residual connection.

Usage

x = rand(Float32, 2, g.num_nodes)
e = rand(Float32, 3, g.num_edges)

l = CGConv((2,3) => 4, tanh)

y = l(g, x, e)    # size: (4, num_nodes)

GraphNeuralNetworks.ChebConv — Type

ChebConv(in => out, k; bias=true, init=glorot_uniform)

Chebyshev spectral graph convolutional layer from paper Convolutional Neural Networks on Graphs with Fast Localized Spectral Filtering.

Implements

\[X' = \sum^{K-1}_{k=0} W^{(k)} Z^{(k)}\]

where $Z^{(k)}$ is the $k$-th term of Chebyshev polynomials, and can be calculated by the following recursive form:

\[Z^{(0)} = X \\ Z^{(1)} = \hat{L} X \\ Z^{(k)} = 2 \hat{L} Z^{(k-1)} - Z^{(k-2)}\]

with $\hat{L}$ the scaled_laplacian.

Arguments

in: The dimension of input features.

out: The dimension of output features.

k: The order of Chebyshev polynomial.

bias: Add learnable bias.

init: Weights' initializer.

GraphNeuralNetworks.EdgeConv — Type

EdgeConv(nn; aggr=max)

Edge convolutional layer from paper Dynamic Graph CNN for Learning on Point Clouds.

Performs the operation

\[\mathbf{x}_i' = \square_{j \in N(i)} nn(\mathbf{x}_i || \mathbf{x}_j - \mathbf{x}_i)\]

where nn generally denotes a learnable function, e.g. a linear layer or a multi-layer perceptron.

Arguments

nn: A (possibly learnable) function acting on edge features.

aggr: Aggregation operator for the incoming messages (e.g. +, *, max, min, and mean).

GraphNeuralNetworks.GATConv — Type

GATConv(in => out, σ=identity;
        heads=1,
        concat=true,
        init=glorot_uniform    
        bias=true, 
        negative_slope=0.2f0)

Graph attentional layer from the paper Graph Attention Networks.

Implements the operation

\[\mathbf{x}_i' = \sum_{j \in N(i)} \alpha_{ij} W \mathbf{x}_j\]

where the attention coefficients $\alpha_{ij}$ are given by

\[\alpha_{ij} = \frac{1}{z_i} \exp(LeakyReLU(\mathbf{a}^T [W \mathbf{x}_i \,\|\, W \mathbf{x}_j]))\]

with $z_i$ a normalization factor.

Arguments

in: The dimension of input features.

out: The dimension of output features.

bias: Learn the additive bias if true.

heads: Number attention heads.

concat: Concatenate layer output or not. If not, layer output is averaged over the heads.

negative_slope: The parameter of LeakyReLU.

GraphNeuralNetworks.GCNConv — Type

GCNConv(in => out, σ=identity; bias=true, init=glorot_uniform, add_self_loops=true)

Graph convolutional layer from paper Semi-supervised Classification with Graph Convolutional Networks.

Performs the operation

\[\mathbf{x}'_i = \sum_{j\in N(i)} \frac{1}{c_{ij}} W \mathbf{x}_j\]

where $c_{ij} = \sqrt{|N(i)||N(j)|}$.

The input to the layer is a node feature array X of size (num_features, num_nodes).

Arguments

in: Number of input features.

out: Number of output features.

σ: Activation function.

bias: Add learnable bias.

init: Weights' initializer.

add_self_loops: Add self loops to the graph before performing the convolution.

GraphNeuralNetworks.GINConv — Type

GINConv(f, ϵ; aggr=+)

Graph Isomorphism convolutional layer from paper How Powerful are Graph Neural Networks?.

Implements the graph convolution

\[\mathbf{x}_i' = f_\Theta\left((1 + \epsilon) \mathbf{x}_i + \sum_{j \in N(i)} \mathbf{x}_j \right)\]

where $f_\Theta$ typically denotes a learnable function, e.g. a linear layer or a multi-layer perceptron.

Arguments

f: A (possibly learnable) function acting on node features.

ϵ: Weighting factor.

GraphNeuralNetworks.GatedGraphConv — Type

GatedGraphConv(out, num_layers; aggr=+, init=glorot_uniform)

Gated graph convolution layer from Gated Graph Sequence Neural Networks.

Implements the recursion

\[\mathbf{h}^{(0)}_i = [\mathbf{x}_i || \mathbf{0}] \\ \mathbf{h}^{(l)}_i = GRU(\mathbf{h}^{(l-1)}_i, \square_{j \in N(i)} W \mathbf{h}^{(l-1)}_j)\]

where $\mathbf{h}^{(l)}_i$ denotes the $l$-th hidden variables passing through GRU. The dimension of input $\mathbf{x}_i$ needs to be less or equal to out.

Arguments

out: The dimension of output features.

num_layers: The number of gated recurrent unit.

aggr: Aggregation operator for the incoming messages (e.g. +, *, max, min, and mean).

init: Weight initialization function.

GraphNeuralNetworks.GraphConv — Type

GraphConv(in => out, σ=identity; aggr=+, bias=true, init=glorot_uniform)

Graph convolution layer from Reference: Weisfeiler and Leman Go Neural: Higher-order Graph Neural Networks.

Performs:

\[\mathbf{x}_i' = W_1 \mathbf{x}_i + \square_{j \in \mathcal{N}(i)} W_2 \mathbf{x}_j\]

where the aggregation type is selected by aggr.

Arguments

in: The dimension of input features.

out: The dimension of output features.

σ: Activation function.

aggr: Aggregation operator for the incoming messages (e.g. +, *, max, min, and mean).

bias: Add learnable bias.

init: Weights' initializer.

GraphNeuralNetworks.NNConv — Type

NNConv(in => out, f, σ=identity; aggr=+, bias=true, init=glorot_uniform)

The continuous kernel-based convolutional operator from the Neural Message Passing for Quantum Chemistry paper. This convolution is also known as the edge-conditioned convolution from the Dynamic Edge-Conditioned Filters in Convolutional Neural Networks on Graphs paper.

Performs the operation

\[\mathbf{x}_i' = W \mathbf{x}_i + \square_{j \in N(i)} f_\Theta(\mathbf{e}_{j\to i})\,\mathbf{x}_j\]

where $f_\Theta$ denotes a learnable function (e.g. a linear layer or a multi-layer perceptron). Given an input of batched edge features e of size (num_edge_features, num_edges), the function f will return an batched matrices array whose size is (out, in, num_edges). For convenience, also functions returning a single (out*in, num_edges) matrix are allowed.

Arguments

in: The dimension of input features.

out: The dimension of output features.

f: A (possibly learnable) function acting on edge features.

aggr: Aggregation operator for the incoming messages (e.g. +, *, max, min, and mean).

σ: Activation function.

bias: Add learnable bias.

init: Weights' initializer.

GraphNeuralNetworks.ResGatedGraphConv — Type

ResGatedGraphConv(in => out, act=identity; init=glorot_uniform, bias=true)

The residual gated graph convolutional operator from the Residual Gated Graph ConvNets paper.

The layer's forward pass is given by

\[\mathbf{x}_i' = act\big(U\mathbf{x}_i + \sum_{j \in N(i)} \eta_{ij} V \mathbf{x}_j\big),\]

where the edge gates $\eta_{ij}$ are given by

\[\eta_{ij} = sigmoid(A\mathbf{x}_i + B\mathbf{x}_j).\]

Arguments

in: The dimension of input features.

out: The dimension of output features.

act: Activation function.

init: Weight matrices' initializing function.

bias: Learn an additive bias if true.

GraphNeuralNetworks.SAGEConv — Type

SAGEConv(in => out, σ=identity; aggr=mean, bias=true, init=glorot_uniform)

GraphSAGE convolution layer from paper Inductive Representation Learning on Large Graphs.

Performs:

\[\mathbf{x}_i' = W \cdot [\mathbf{x}_i \,\|\, \square_{j \in \mathcal{N}(i)} \mathbf{x}_j]\]

where the aggregation type is selected by aggr.

Arguments

in: The dimension of input features.

out: The dimension of output features.

σ: Activation function.

aggr: Aggregation operator for the incoming messages (e.g. +, *, max, min, and mean).

bias: Add learnable bias.

init: Weights' initializer.