Convolutional Layers

GraphNeuralNetworks.ChebConv — Type

ChebConv(in=>out, k; bias=true, init=glorot_uniform)

Chebyshev spectral graph convolutional layer from paper Convolutional Neural Networks on Graphs with Fast Localized Spectral Filtering.

Implements

\[X' = \sum^{K-1}_{k=0} W^{(k)} Z^{(k)}\]

where $Z^{(k)}$ is the $k$-th term of Chebyshev polynomials, and can be calculated by the following recursive form:

\[Z^{(0)} = X \\ Z^{(1)} = \hat{L} X \\ Z^{(k)} = 2 \hat{L} Z^{(k-1)} - Z^{(k-2)}\]

with $\hat{L}$ the scaled_laplacian.

Arguments

in: The dimension of input features.

out: The dimension of output features.

k: The order of Chebyshev polynomial.

bias: Add learnable bias.

init: Weights' initializer.

GraphNeuralNetworks.EdgeConv — Type

EdgeConv(f; aggr=max)

Edge convolutional layer from paper Dynamic Graph CNN for Learning on Point Clouds.

Performs the operation

\[\mathbf{x}_i' = \square_{j \in N(i)} f(\mathbf{x}_i || \mathbf{x}_j - \mathbf{x}_i)\]

where f typically denotes a learnable function, e.g. a linear layer or a multi-layer perceptron.

Arguments

f: A (possibly learnable) function acting on edge features.

aggr: Aggregation operator for the incoming messages (e.g. +, *, max, min, and mean).

GraphNeuralNetworks.GATConv — Type

GATConv(in => out, , σ=identity;
        heads=1,
        concat=true,
        init=glorot_uniform    
        bias=true, 
        negative_slope=0.2f0)

Graph attentional layer from the paper Graph Attention Networks.

Implements the operation

\[\mathbf{x}_i' = \sum_{j \in N(i)} \alpha_{ij} W \mathbf{x}_j\]

where the attention coefficient $\alpha_{ij}$ is given by

\[\alpha_{ij} = \frac{1}{z_i} \exp(LeakyReLU(\mathbf{a}^T [W \mathbf{x}_i || W \mathbf{x}_j]))\]

with $z_i$ a normalization factor.

Arguments

in: The dimension of input features.

out: The dimension of output features.

bias::Bool: Keyword argument, whether to learn the additive bias.

heads: Number attention heads

concat: Concatenate layer output or not. If not, layer output is averaged over the heads.

negative_slope::Real: Keyword argument, the parameter of LeakyReLU.

GraphNeuralNetworks.GCNConv — Type

GCNConv(in => out, σ=identity; bias=true, init=glorot_uniform)

Graph convolutional layer from paper Semi-supervised Classification with Graph Convolutional Networks.

Performs the operation

\[\mathbf{x}'_i = \sum_{j\in N(i)} \frac{1}{c_{ij}} W \mathbf{x}_j\]

where $c_{ij} = \sqrt{N(i)\,N(j)}$.

The input to the layer is a node feature array X of size (num_features, num_nodes).

Arguments

in: Number of input features.

out: Number of output features.

σ: Activation function.

bias: Add learnable bias.

init: Weights' initializer.

GraphNeuralNetworks.GINConv — Type

GINConv(f; eps = 0f0)

Graph Isomorphism convolutional layer from paper How Powerful are Graph Neural Networks?

\[\mathbf{x}_i' = f\left((1 + \epsilon) \mathbf{x}_i + \sum_{j \in N(i)} \mathbf{x}_j \right)\]

where f typically denotes a learnable function, e.g. a linear layer or a multi-layer perceptron.

Arguments

f: A (possibly learnable) function acting on node features.

eps: Weighting factor.

GraphNeuralNetworks.GatedGraphConv — Type

GatedGraphConv(out, num_layers; aggr=+, init=glorot_uniform)

Gated graph convolution layer from Gated Graph Sequence Neural Networks.

Implements the recursion

\[\mathbf{h}^{(0)}_i = \mathbf{x}_i || \mathbf{0} \\ \mathbf{h}^{(l)}_i = GRU(\mathbf{h}^{(l-1)}_i, \square_{j \in N(i)} W \mathbf{h}^{(l-1)}_j)\]

where $\mathbf{h}^{(l)}_i$ denotes the $l$-th hidden variables passing through GRU. The dimension of input $\mathbf{x}_i$ needs to be less or equal to out.

Arguments

out: The dimension of output features.

num_layers: The number of gated recurrent unit.

aggr: Aggregation operator for the incoming messages (e.g. +, *, max, min, and mean).

init: Weight initialization function.

GraphNeuralNetworks.GraphConv — Type

GraphConv(in => out, σ=identity, aggr=+; bias=true, init=glorot_uniform)

Graph convolution layer from Reference: Weisfeiler and Leman Go Neural: Higher-order Graph Neural Networks.

Performs:

\[\mathbf{x}_i' = W^1 \mathbf{x}_i + \square_{j \in \mathcal{N}(i)} W^2 \mathbf{x}_j)\]

where the aggregation type is selected by aggr.

Arguments

in: The dimension of input features.

out: The dimension of output features.

σ: Activation function.

aggr: Aggregation operator for the incoming messages (e.g. +, *, max, min, and mean).

bias: Add learnable bias.

init: Weights' initializer.