Graphs

GraphNeuralNetworks.GNNGraph — Type

GNNGraph(data; [graph_type, nf, ef, gf, num_nodes, graph_indicator, dir])
GNNGraph(g::GNNGraph; [nf, ef, gf])

A type representing a graph structure and storing also arrays that contain features associated to nodes, edges, and the whole graph.

A GNNGraph can be constructed out of different objects data representing the connections inside the graph, while the internal representation type is governed by graph_type. When constructed from another graph g, the internal graph representation is preserved and shared.

A GNNGraph can also represent multiple graphs batched togheter (see Flux.batch or SparseArrays.blockdiag). The field g.graph_indicator contains the graph membership of each node.

A GNNGraph is a LightGraphs' AbstractGraph, therefore any functionality from the LightGraphs' graph library can be used on it.

Arguments

data: Some data representing the graph topology. Possible type are

An adjacency matrix

An adjacency list.

A tuple containing the source and target vectors (COO representation)

A LightGraphs' graph.

graph_type: A keyword argument that specifies the underlying representation used by the GNNGraph. Currently supported values are

:coo. Graph represented as a tuple (source, target), such that the k-th edge connects the node source[k] to node target[k]. Optionally, also edge weights can be given: (source, target, weights).

:sparse. A sparse adjacency matrix representation.

:dense. A dense adjacency matrix representation.

Default :coo.

dir. The assumed edge direction when given adjacency matrix or adjacency list input data g. Possible values are :out and :in. Default :out.

num_nodes. The number of nodes. If not specified, inferred from g. Default nothing.

graph_indicator. For batched graphs, a vector containeing the graph assigment of each node. Default nothing.

nf: Node features. Either nothing, or an array whose last dimension has size num_nodes. Default nothing.

ef: Edge features. Either nothing, or an array whose last dimension has size num_edges. Default nothing.

gf: Global features. Default nothing.

Usage.

using Flux, GraphNeuralNetworks

# Construct from adjacency list representation
data = [[2,3], [1,4,5], [1], [2,5], [2,4]]
g = GNNGraph(data)

# Number of nodes and edges
g.num_nodes  # 5
g.num_edges  # 10 

# Same graph in COO representation
s = [1,1,2,2,2,3,4,4,5,5]
t = [2,3,1,4,5,3,2,5,2,4]
g = GNNGraph(s, t)

# From a LightGraphs' graph
g = GNNGraph(erdos_renyi(100, 20))

# Copy graph while also adding node features
g = GNNGraph(g, nf=rand(100, 5))

# Send to gpu
g = g |> gpu

# Collect edges' source and target nodes.
# Both source and target are vectors of length num_edges
source, target = edge_index(g)

See also graph, edge_index, node_feature, edge_feature, and global_feature

GraphNeuralNetworks.add_self_loops — Method

add_self_loops(g::GNNGraph)

Return a graph with the same features as g but also adding edges connecting the nodes to themselves.

Nodes with already existing self-loops will obtain a second self-loop.

GraphNeuralNetworks.edge_feature — Method

edge_feature(g::GNNGraph)

Return the edge features of g.

GraphNeuralNetworks.edge_index — Method

edge_index(g::GNNGraph)

Return a tuple containing two vectors, respectively storing the source and target nodes for each edges in g.

s, t = edge_index(g)

GraphNeuralNetworks.global_feature — Method

global_feature(g::GNNGraph)

Return the global features of g.

GraphNeuralNetworks.node_feature — Method

node_feature(g::GNNGraph)

Return the node features of g.

GraphNeuralNetworks.normalized_laplacian — Function

normalized_laplacian(g, T=Float32; add_self_loops=false, dir=:out)

Normalized Laplacian matrix of graph g.

Arguments

g: A GNNGraph.

T: result element type.

add_self_loops: add self-loops while calculating the matrix.

dir: the edge directionality considered (:out, :in, :both).

GraphNeuralNetworks.scaled_laplacian — Function

scaled_laplacian(g, T=Float32; dir=:out)

Scaled Laplacian matrix of graph g, defined as $\hat{L} = \frac{2}{\lambda_{max}} L - I$ where $L$ is the normalized Laplacian matrix.

Arguments

g: A GNNGraph.

T: result element type.

dir: the edge directionality considered (:out, :in, :both).

GraphNeuralNetworks.subgraph — Method

subgraph(g::GNNGraph, i)

Return the subgraph of g induced by those nodes v for which g.graph_indicator[v] ∈ i. In other words, it extract the component graphs from a batched graph.

It also returns a vector nodes mapping the new nodes to the old ones. The node i in the subgraph corresponds to the node nodes[i] in g.