GNNGraph

Documentation page for the graph type GNNGraph provided by GraphNeuralNetworks.jl and related methods.

GNNGraph

Index

GNNGraph type

Query

Transform

Generate

Besides the methods documented here, one can rely on the large set of functionalities given by Graphs.jl since GNNGraph inherits from Graphs.AbstractGraph.

Index

GraphNeuralNetworks.GNNGraphs.GNNGraph

Flux.batch

Flux.unbatch

GraphNeuralNetworks.GNNGraphs.add_edges

GraphNeuralNetworks.GNNGraphs.add_nodes

GraphNeuralNetworks.GNNGraphs.add_self_loops

GraphNeuralNetworks.GNNGraphs.adjacency_list

GraphNeuralNetworks.GNNGraphs.edge_index

GraphNeuralNetworks.GNNGraphs.getgraph

GraphNeuralNetworks.GNNGraphs.graph_indicator

GraphNeuralNetworks.GNNGraphs.negative_sample

GraphNeuralNetworks.GNNGraphs.normalized_laplacian

GraphNeuralNetworks.GNNGraphs.rand_graph

GraphNeuralNetworks.GNNGraphs.scaled_laplacian

Graphs.LinAlg.adjacency_matrix

Graphs.degree

Graphs.inneighbors

Graphs.outneighbors

SparseArrays.blockdiag

GNNGraph type

GraphNeuralNetworks.GNNGraphs.GNNGraph — Type

GNNGraph(data; [graph_type, ndata, edata, gdata, num_nodes, graph_indicator, dir])
GNNGraph(g::GNNGraph; [ndata, edata, gdata])

A type representing a graph structure that also stores feature arrays associated to nodes, edges, and the graph itself.

A GNNGraph can be constructed out of different data objects expressing the connections inside the graph. The internal representation type is determined by graph_type.

When constructed from another GNNGraph, the internal graph representation is preserved and shared. The node/edge/graph features are retained as well, unless explicitely set by the keyword arguments ndata, edata, and gdata.

A GNNGraph can also represent multiple graphs batched togheter (see Flux.batch or SparseArrays.blockdiag). The field g.graph_indicator contains the graph membership of each node.

GNNGraphs are always directed graphs, therefore each edge is defined by a source node and a target node (see edge_index). Self loops (edges connecting a node to itself) and multiple edges (more than one edge between the same pair of nodes) are supported.

A GNNGraph is a Graphs.jl's AbstractGraph, therefore it supports most functionality from that library.

Arguments

data: Some data representing the graph topology. Possible type are

An adjacency matrix

An adjacency list.

A tuple containing the source and target vectors (COO representation)

A Graphs' graph.

graph_type: A keyword argument that specifies the underlying representation used by the GNNGraph. Currently supported values are

:coo. Graph represented as a tuple (source, target), such that the k-th edge connects the node source[k] to node target[k]. Optionally, also edge weights can be given: (source, target, weights).

:sparse. A sparse adjacency matrix representation.

:dense. A dense adjacency matrix representation.

Defaults to :coo, currently the most supported type.

dir: The assumed edge direction when given adjacency matrix or adjacency list input data g. Possible values are :out and :in. Default :out.

num_nodes: The number of nodes. If not specified, inferred from g. Default nothing.

graph_indicator: For batched graphs, a vector containing the graph assigment of each node. Default nothing.

ndata: Node features. A named tuple of arrays whose last dimension has size num_nodes.

edata: Edge features. A named tuple of arrays whose last dimension has size num_edges.

gdata: Graph features. A named tuple of arrays whose last dimension has size num_graphs.

Usage.

using Flux, GraphNeuralNetworks

# Construct from adjacency list representation
data = [[2,3], [1,4,5], [1], [2,5], [2,4]]
g = GNNGraph(data)

# Number of nodes, edges, and batched graphs
g.num_nodes  # 5
g.num_edges  # 10 
g.num_graphs # 1 

# Same graph in COO representation
s = [1,1,2,2,2,3,4,4,5,5]
t = [2,3,1,4,5,3,2,5,2,4]
g = GNNGraph(s, t)

# From a Graphs' graph
g = GNNGraph(erdos_renyi(100, 20))

# Add 2 node feature arrays
g = GNNGraph(g, ndata = (x=rand(100, g.num_nodes), y=rand(g.num_nodes)))

# Add node features and edge features with default names `x` and `e` 
g = GNNGraph(g, ndata = rand(100, g.num_nodes), edata = rand(16, g.num_edges))

g.ndata.x
g.ndata.e

# Send to gpu
g = g |> gpu

# Collect edges' source and target nodes.
# Both source and target are vectors of length num_edges
source, target = edge_index(g)

Query

GraphNeuralNetworks.GNNGraphs.adjacency_list — Method

adjacency_list(g; dir=:out)

Return the adjacency list representation (a vector of vectors) of the graph g.

Calling a the adjacency list, if dir=:out than a[i] will contain the neighbors of node i through outgoing edges. If dir=:in, it will contain neighbors from incoming edges instead.

GraphNeuralNetworks.GNNGraphs.edge_index — Method

edge_index(g::GNNGraph)

Return a tuple containing two vectors, respectively storing the source and target nodes for each edges in g.

s, t = edge_index(g)

GraphNeuralNetworks.GNNGraphs.graph_indicator — Method

graph_indicator(g)

Return a vector containing the graph membership (an integer from 1 to g.num_graphs) of each node in the graph.

GraphNeuralNetworks.GNNGraphs.normalized_laplacian — Function

normalized_laplacian(g, T=Float32; add_self_loops=false, dir=:out)

Normalized Laplacian matrix of graph g.

Arguments

g: A GNNGraph.

T: result element type.

add_self_loops: add self-loops while calculating the matrix.

dir: the edge directionality considered (:out, :in, :both).

GraphNeuralNetworks.GNNGraphs.scaled_laplacian — Function

scaled_laplacian(g, T=Float32; dir=:out)

Scaled Laplacian matrix of graph g, defined as $\hat{L} = \frac{2}{\lambda_{max}} L - I$ where $L$ is the normalized Laplacian matrix.

Arguments

g: A GNNGraph.

T: result element type.

dir: the edge directionality considered (:out, :in, :both).

Graphs.LinAlg.adjacency_matrix — Function

adjacency_matrix(g[, T=Int; dir=:out])

Return a sparse adjacency matrix for a graph, indexed by [u, v] vertices. Non-zero values indicate an edge from u to v. Users may override the default data type (Int) and specify an optional direction.

Optional Arguments

dir=:out: :in, :out, or :both are currently supported.

Implementation Notes

This function is optimized for speed and directly manipulates CSC sparse matrix fields.

Graphs.degree — Function

degree(g[, v])

Return a vector corresponding to the number of edges which start or end at each vertex in graph g. If v is specified, only return degrees for vertices in v. For directed graphs, this value equals the incoming plus outgoing edges. For undirected graphs, it equals the connected edges.

Examples

julia> using Graphs

julia> g = DiGraph(3);

julia> add_edge!(g, 2, 3);

julia> add_edge!(g, 3, 1);

julia> degree(g)
3-element Array{Int64,1}:
 1
 1
 2

Graphs.outneighbors — Function

outneighbors(g, v)

Return a list of all neighbors connected to vertex v by an outgoing edge.

Implementation Notes

Returns a reference to the current graph's internal structures, not a copy. Do not modify result. If the graph is modified, the behavior is undefined: the array behind this reference may be modified too, but this is not guaranteed.

Examples

julia> g = SimpleDiGraph([0 1 0 0 0; 0 0 1 0 0; 1 0 0 1 0; 0 0 0 0 1; 0 0 0 1 0]);

julia> outneighbors(g, 4)
1-element Array{Int64,1}:
 5

Graphs.inneighbors — Function

inneighbors(g, v)

Return a list of all neighbors connected to vertex v by an incoming edge.

Implementation Notes

Returns a reference to the current graph's internal structures, not a copy. Do not modify result. If the graph is modified, the behavior is undefined: the array behind this reference may be modified too, but this is not guaranteed.

Examples

julia> g = SimpleDiGraph([0 1 0 0 0; 0 0 1 0 0; 1 0 0 1 0; 0 0 0 0 1; 0 0 0 1 0]);

julia> inneighbors(g, 4)
2-element Array{Int64,1}:
 3
 5

Transform

Flux.batch — Method

batch(gs::Vector{<:GNNGraph})

Batch together multiple GNNGraphs into a single one containing the total number of original nodes and edges.

Equivalent to SparseArrays.blockdiag. See also Flux.unbatch.

Usage

julia> g1 = rand_graph(4, 6, ndata=ones(8, 4))
GNNGraph:
    num_nodes = 4
    num_edges = 6
    num_graphs = 1
    ndata:
        x => (8, 4)
    edata:
    gdata:

julia> g2 = rand_graph(7, 4, ndata=zeros(8, 7))
GNNGraph:
    num_nodes = 7
    num_edges = 4
    num_graphs = 1
    ndata:
        x => (8, 7)
    edata:
    gdata:

julia> g12 = Flux.batch([g1, g2])
GNNGraph:
    num_nodes = 11
    num_edges = 10
    num_graphs = 2
    ndata:
        x => (8, 11)
    edata:
    gdata:

julia> g12.ndata.x
8×11 Matrix{Float64}:
 1.0  1.0  1.0  1.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0
 1.0  1.0  1.0  1.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0
 1.0  1.0  1.0  1.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0
 1.0  1.0  1.0  1.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0
 1.0  1.0  1.0  1.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0
 1.0  1.0  1.0  1.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0
 1.0  1.0  1.0  1.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0
 1.0  1.0  1.0  1.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0

Flux.unbatch — Method

unbatch(g::GNNGraph)

Opposite of the Flux.batch operation, returns an array of the individual graphs batched together in g.

See also Flux.batch and getgraph.

Usage

julia> gbatched = Flux.batch([rand_graph(5, 6), rand_graph(10, 8), rand_graph(4,2)])
GNNGraph:
    num_nodes = 19
    num_edges = 16
    num_graphs = 3
    ndata:
    edata:
    gdata:

julia> Flux.unbatch(gbatched)
3-element Vector{GNNGraph{Tuple{Vector{Int64}, Vector{Int64}, Nothing}}}:
 GNNGraph:
    num_nodes = 5
    num_edges = 6
    num_graphs = 1
    ndata:
    edata:
    gdata:

 GNNGraph:
    num_nodes = 10
    num_edges = 8
    num_graphs = 1
    ndata:
    edata:
    gdata:

 GNNGraph:
    num_nodes = 4
    num_edges = 2
    num_graphs = 1
    ndata:
    edata:
    gdata:

GraphNeuralNetworks.GNNGraphs.add_edges — Method

add_edges(g::GNNGraph, s::AbstractVector, t::AbstractVector; [edata])

Add to graph g the edges with source nodes s and target nodes t.

GraphNeuralNetworks.GNNGraphs.add_nodes — Method

add_nodes(g::GNNGraph, n; [ndata])

Add n new nodes to graph g. In the new graph, these nodes will have indexes from g.num_nodes + 1 to g.num_nodes + n.

GraphNeuralNetworks.GNNGraphs.add_self_loops — Method

add_self_loops(g::GNNGraph)

Return a graph with the same features as g but also adding edges connecting the nodes to themselves.

Nodes with already existing self-loops will obtain a second self-loop.

GraphNeuralNetworks.GNNGraphs.getgraph — Method

getgraph(g::GNNGraph, i; nmap=false)

Return the subgraph of g induced by those nodes j for which g.graph_indicator[j] == i or, if i is a collection, g.graph_indicator[j] ∈ i. In other words, it extract the component graphs from a batched graph.

If nmap=true, return also a vector v mapping the new nodes to the old ones. The node i in the subgraph will correspond to the node v[i] in g.

GraphNeuralNetworks.GNNGraphs.negative_sample — Method

negative_sample(g::GNNGraph; num_neg_edges=g.num_edges)

Return a graph containing random negative edges (i.e. non-edges) from graph g.

SparseArrays.blockdiag — Method

blockdiag(xs::GNNGraph...)

Equivalent to Flux.batch.

Generate

GraphNeuralNetworks.GNNGraphs.rand_graph — Method

rand_graph(n, m; bidirected=true, seed=-1, kws...)

Generate a random (Erdós-Renyi) GNNGraph with n nodes and m edges.

If bidirected=true the reverse edge of each edge will be present. If bidirected=false instead, m unrelated edges are generated. In any case, the output graph will contain no self-loops or multi-edges.

Use a seed > 0 for reproducibility.

Additional keyword arguments will be passed to the GNNGraph constructor.

Usage

julia> g = rand_graph(5, 4, bidirected=false)
GNNGraph:
    num_nodes = 5
    num_edges = 4
    num_graphs = 1
    ndata:
    edata:
    gdata:

julia> edge_index(g)
([1, 3, 3, 4], [5, 4, 5, 2])

# In the bidirected case, edge data will be duplicated on the reverse edges if needed.
julia> g = rand_graph(5, 4, edata=rand(16, 2))
GNNGraph:
    num_nodes = 5
    num_edges = 4
    num_graphs = 1
    ndata:
    edata:
        e => (16, 4)
    gdata:

# Each edge has a reverse
julia> edge_index(g)
([1, 3, 3, 4], [3, 4, 1, 3])