Message Passing

Index

GraphNeuralNetworks.compute_message

GraphNeuralNetworks.propagate

GraphNeuralNetworks.update_edge

GraphNeuralNetworks.update_node

Docs

GraphNeuralNetworks.compute_message — Function

compute_message(l, x_i, x_j, [e_ij])

Message function for the message-passing scheme started by propagate. Returns the message from node j to node i . In the message-passing scheme, the incoming messages from the neighborhood of i will later be aggregated in order to update (see update_node) the features of node i.

The function operates on batches of edges, therefore x_i, x_j, and e_ij are tensors whose last dimension is the batch size, or can be tuple/namedtuples of such tensors, according to the input to propagate.

By default, the function returns x_j. Custom layer should specialize this method with the desired behavior.

Arguments

l: A gnn layer.

x_i: Features of the central node i.

x_j: Features of the neighbor j of node i.

e_ij: Features of edge (i,j).

See also update_node and propagate.

GraphNeuralNetworks.update_node — Function

update_node(l, m̄, x)

Node update function for the GNN layer l, returning a new set of node features x′ based on old features x and the aggregated message m̄ from the neighborhood.

The input m̄ is an array, a tuple or a named tuple, reflecting the output of compute_message.

By default, the function returns m̄. Custom layers should specialize this method with the desired behavior.

See also compute_message, update_edge, and propagate.

GraphNeuralNetworks.update_edge — Function

update_edge(l, m, e)

Edge update function for the GNN layer l, returning a new set of edge features e′ based on old features e and the newly computed messages m from the compute_message function.

By default, the function returns e. Custom layers should specialize this method with the desired behavior.

See also compute_message, update_node, and propagate.

GraphNeuralNetworks.propagate — Function

propagate(l, g, aggr, [x, e]) -> x′, e′
propagate(l, g, aggr) -> g′

Performs the message-passing for GNN layer l on graph g . Returns updated node and edge features x and e.

In case no input and edge features are given as input, extracts them from g and returns the same graph with updated feautres.

The computational steps are the following:

m = compute_batch_message(l, g, x, e)  # calls `compute_message`
m̄ = aggregate_neighbors(g, aggr, m)
x′ = update_node(l, m̄, x)
e′ = update_edge(l, m, e)

Custom layers typically define their own update_node and compute_message functions, then call this method in the forward pass:

Usage example

using GraphNeuralNetworks, Flux

struct GNNConv <: GNNLayer
    W
    b
    σ
end

Flux.@functor GNNConv

function GNNConv(ch::Pair{Int,Int}, σ=identity;
                 init=glorot_uniform, bias::Bool=true)
    in, out = ch
    W = init(out, in)
    b = Flux.create_bias(W, bias, out)
    GNNConv(W, b, σ, aggr)
end

compute_message(l::GNNConv, x_i, x_j, e_ij) = l.W * x_j
update_node(l::GNNConv, m̄, x) = l.σ.(m̄ .+ l.bias)

function (l::GNNConv)(g::GNNGraph, x::AbstractMatrix)
    x, _ = propagate(l, g, +, x)
    return x
end

See also compute_message and update_node.