Message Passing

GraphNeuralNetworks.message — Function

message(mp, x_i, x_j, [e_ij, u])

Message function for the message-passing scheme, returning the message from node j to node i . In the message-passing scheme. the incoming messages from the neighborhood of i will later be aggregated in order to update the features of node i.

By default, the function returns x_j. Custom layer should specialize this method with the desired behavior.

Arguments

mp: A gnn layer.

x_i: Features of the central node i.

x_j: Features of the neighbor j of node i.

e_ij: Features of edge (i, j).

u: Global features.

See also update and propagate.

GraphNeuralNetworks.update — Function

update(mp, m̄, x, [u])

Update function for the message-passing scheme, returning a new set of node features x′ based on old features x and the incoming message from the neighborhood aggregation m̄.

By default, the function returns m̄. Custom layers should specialize this method with the desired behavior.

Arguments

mp: A gnn layer.

m̄: Aggregated edge messages from the message function.

x: Node features to be updated.

u: Global features.

See also message and propagate.

GraphNeuralNetworks.propagate — Function

propagate(mp, g::GNNGraph, aggr)
propagate(mp, g::GNNGraph, E, X, u, aggr)

Perform the sequence of operation implementing the message-passing scheme and updating node, edge, and global features X, E, and u respectively.

The computation involved is the following:

M = compute_batch_message(mp, g, E, X, u) 
E = update_edge(mp, M, E, u)
M̄ = aggregate_neighbors(mp, aggr, g, M)
X = update(mp, M̄, X, u)
u = update_global(mp, E, X, u)

Custom layers typically define their own update and message function, then call this method in the forward pass:

function (l::MyLayer)(g, X)
    ... some prepocessing if needed ...
    E = nothing
    u = nothing
    propagate(l, g, E, X, u, +)
end

See also message and update.