Pooling Layers

GraphNeuralNetworks.GlobalPool — Type

GlobalPool(aggr)

Global pooling layer for graph neural networks. Takes a graph and feature nodes as inputs and performs the operation

\[\mathbf{u}_V = \box_{i \in V} \mathbf{x}_i ```` where ``V`` is the set of nodes of the input graph and the type of aggregation represented by `\box` is selected by the `aggr` argument. Commonly used aggregations are are `mean`, `max`, and `+`. \]

julia using GraphNeuralNetworks, LightGraphs

pool = GlobalPool(mean)

g = GNNGraph(randomregulargraph(10, 4)) X = rand(32, 10) pool(g, X) # => 32x1 matrix ```

GraphNeuralNetworks.TopKPool — Type

TopKPool(adj, k, in_channel)

Top-k pooling layer.

Arguments

adj: Adjacency matrix of a graph.

k: Top-k nodes are selected to pool together.

in_channel: The dimension of input channel.