Temporal Graph-Convolutional Layers

Convolutions for time-varying graphs (temporal graphs) such as the TemporalSnapshotsGNNGraph.

GraphNeuralNetworks.TGCN — Type

TGCN(din => dout; [bias, init, add_self_loops])

Temporal Graph Convolutional Network (T-GCN) recurrent layer from the paper T-GCN: A Temporal Graph Convolutional Network for Traffic Prediction.

Performs a layer of GCNConv to model spatial dependencies, followed by a Gated Recurrent Unit (GRU) cell to model temporal dependencies.

Arguments

din: Number of input features.

dout: Number of output features.

bias: Add learnable bias. Default true.

init: Convolution's weights initializer. Default glorot_uniform.

add_self_loops: Add self loops to the graph before performing the convolution. Default false.

Forward

tgcn(g::GNNGraph, x, [h])

g: The input graph.

x: The input to the TGCN. It should be a matrix size din x timesteps or an array of size din x timesteps x num_nodes.

h: The initial hidden state of the GRU cell. If given, it is a vector of size out or a matrix of size dout x num_nodes. If not provided, it is assumed to be a vector of zeros.

Examples

julia> din, dout = 2, 3;

julia> tgcn = TGCN(din => dout)
TGCN(
  TGCNCell(
    GCNConv(2 => 3, σ),                 # 9 parameters
    GRUCell(3 => 3),                    # 63 parameters
  ),
)                   # Total: 5 arrays, 72 parameters, 560 bytes.

julia> num_nodes = 5; num_edges = 10; timesteps = 4;

julia> g = rand_graph(num_nodes, num_edges);

julia> x = rand(Float32, din, timesteps, num_nodes);

julia> tgcn(g, x) |> size
(3, 4, 5)