Utility Functions

Index

GraphNeuralNetworks.broadcast_edges

GraphNeuralNetworks.broadcast_nodes

GraphNeuralNetworks.reduce_edges

GraphNeuralNetworks.reduce_nodes

GraphNeuralNetworks.softmax_edges

GraphNeuralNetworks.softmax_nodes

NNlib.gather

NNlib.gather!

NNlib.scatter

NNlib.scatter!

Docs

Graph-wise operations

GraphNeuralNetworks.reduce_nodes — Function

reduce_nodes(aggr, g, x)

For a batched graph g, return the graph-wise aggregation of the node features x. The aggregation operator aggr can be +, mean, max, or min. The returned array will have last dimension g.num_graphs.

GraphNeuralNetworks.reduce_edges — Function

reduce_edges(aggr, g, e)

For a batched graph g, return the graph-wise aggregation of the edge features e. The aggregation operator aggr can be +, mean, max, or min. The returned array will have last dimension g.num_graphs.

GraphNeuralNetworks.softmax_nodes — Function

softmax_nodes(g, x)

Graph-wise softmax of the node features x.

GraphNeuralNetworks.softmax_edges — Function

softmax_edges(g, e)

Graph-wise softmax of the edge features e.

GraphNeuralNetworks.broadcast_nodes — Function

broadcast_nodes(g, x)

Graph-wise broadcast array x of size (*, g.num_graphs) to size (*, g.num_nodes).

GraphNeuralNetworks.broadcast_edges — Function

broadcast_edges(g, x)

Graph-wise broadcast array x of size (*, g.num_graphs) to size (*, g.num_edges).

NNlib

Primitive functions implemented in NNlib.jl.

NNlib.gather! — Function

gather!(dst, src, idx)

Reverse operation of scatter!. Gathers data from source src and writes it in destination dst according to the index array idx. For each k in CartesianIndices(idx), assign values to dst according to

dst[:, ... , k] .= src[:, ... , idx[k]...]

Notice that if idx is a vector containing integers, and both dst and src are matrices, previous expression simplifies to

dst[:, k] .= src[:, idx[k]]

and k will run over 1:length(idx).

The elements of idx can be integers or integer tuples and may be repeated. A single src column can end up being copied into zero, one, or multiple dst columns.

See gather for an allocating version.

NNlib.gather — Function

gather(src, idx) -> dst

Reverse operation of scatter. Gathers data from source src and writes it in a destination dst according to the index array idx. For each k in CartesianIndices(idx), assign values to dst according to

dst[:, ... , k] .= src[:, ... , idx[k]...]

Notice that if idx is a vector containing integers and src is a matrix, previous expression simplifies to

dst[:, k] .= src[:, idx[k]]

and k will run over 1:length(idx).

The elements of idx can be integers or integer tuples and may be repeated. A single src column can end up being copied into zero, one, or multiple dst columns.

See gather! for an in-place version.

NNlib.scatter! — Function

scatter!(op, dst, src, idx)

Scatter operation, which scatters data in src and assigns to dst according to idx. A binary reduction operator op is applied during the scatter. For each index k in idx, accumulates values in dst according to

dst[:, ..., idx[k]...] = (op).(dst[:, ..., idx[k]...], src[:, ..., k...])

Arguments

op: Operations to be applied on dst and src, e.g. +, -, *, /, max, min and mean.

dst: The destination for src to aggregate to. This argument will be mutated.

src: The source data for aggregating.

idx: The mapping for aggregation from source (index) to destination (value). The idx array can contain either integers or tuples.

NNlib.scatter — Function

scatter(op, src, idx; [init, dstsize])

Scatter operation allocating a destination array dst and calling scatter!(op, dst, src, idx) on it.

If init is provided, it is used to initialize the content of dst. Otherwise, the init values is inferred from the reduction operator op for some common operators (e.g. init = 0 for op = +).

If dstsize is provided, it will be used to define the size of destination array, otherwise it will be inferred by src and idx.

See scatter! for the details.